+++ tests/unit_tests/Commands/Site/DeleteCommandTest.php
@@ -3,8 +3,9 @@
 namespace Pantheon\Terminus\UnitTests\Commands\Site;
 
 use Pantheon\Terminus\Commands\Site\DeleteCommand;
+use Pantheon\Terminus\Models\Workflow;
 use Pantheon\Terminus\UnitTests\Commands\CommandTestCase;
-use Symfony\Component\Console\Input\Input;
+use Pantheon\Terminus\UnitTests\Commands\WorkflowProgressTrait;
 
 /**
  * Class DeleteCommandTest
@@ -13,12 +14,38 @@ use Symfony\Component\Console\Input\Input;
  */
 class DeleteCommandTest extends CommandTestCase
 {
+    use WorkflowProgressTrait;
+
+    /**
+     * @var string
+     */
+    protected $message;
+    /**
+     * @var string
+     */
+    protected $site_name;
+    /**
+     * @var Workflow
+     */
+    protected $workflow;
+
     /**
      * @inheritdoc
      */
     protected function setup()
     {
         parent::setUp();
+
+        $this->workflow = $this->getMockBuilder(Workflow::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+        $this->message = 'Deleted {site} from Pantheon';
+
+        $this->site_name = 'my-site';
+        $this->site->expects($this->once())
+            ->method('getName')
+            ->with()
+            ->willReturn($this->site_name);
 
         $this->command = new DeleteCommand($this->getConfig());
         $this->command->setSites($this->sites);
@@ -31,20 +58,71 @@ class DeleteCommandTest extends CommandTestCase
      */
     public function testDelete()
     {
-        $site_name = 'my-site';
-
         $this->expectConfirmation();
+        $this->expectWorkflowProcessing();
         $this->site->expects($this->once())
             ->method('delete')
-            ->with();
+            ->with()
+            ->willReturn($this->workflow);
+        $this->workflow->expects($this->once())
+            ->method('getMessage')
+            ->with()
+            ->willReturn($this->message);
         $this->logger->expects($this->once())
             ->method('log')->with(
                 $this->equalTo('notice'),
-                $this->equalTo('Deleted {site} from Pantheon'),
-                $this->equalTo(['site' => $site_name,])
+                $this->equalTo($this->message),
+                $this->equalTo(['site' => $this->site_name,])
             );
 
-        $out = $this->command->delete($site_name);
+        $out = $this->command->delete($this->site_name);
+        $this->assertNull($out);
+    }
+
+    /**
+     * Exercises the site:delete command when the workflow ends with a 404 error
+     */
+    public function testDelete404()
+    {
+        $this->expectConfirmation();
+        $this->expectWorkflowProcessing();
+        $this->site->expects($this->once())
+            ->method('delete')
+            ->with()
+            ->willReturn($this->workflow);
+        $this->progress_bar->method('cycle')
+            ->will($this->throwException(new \Exception($this->message, 404)));
+        $this->workflow->expects($this->never())
+            ->method('getMessage');
+        $this->logger->expects($this->once())
+            ->method('log')->with(
+                $this->equalTo('notice'),
+                $this->equalTo($this->message),
+                $this->equalTo(['site' => $this->site_name,])
+            );
+
+        $out = $this->command->delete($this->site_name);
+        $this->assertNull($out);
+    }
+
+    /**
+     * Exercises the site:delete command when the workflow returns a non-404 error
+     */
+    public function testDeleteErrs()
+    {
+        $this->expectConfirmation();
+        $this->expectWorkflowProcessing();
+        $this->site->expects($this->once())
+            ->method('delete')
+            ->with()
+            ->willReturn($this->workflow);
+        $this->progress_bar->method('cycle')
+            ->will($this->throwException(new \Exception($this->message, 403)));
+        $this->setExpectedException(\Exception::class, $this->message);
+        $this->logger->expects($this->never())
+            ->method('log');
+
+        $out = $this->command->delete($this->site_name);
         $this->assertNull($out);
     }
 
@@ -55,15 +133,13 @@ class DeleteCommandTest extends CommandTestCase
      */
     public function testDeleteConfirmationDecline()
     {
-        $site_name = 'my-site';
-
         $this->expectConfirmation(false);
         $this->site->expects($this->never())
             ->method('delete');
         $this->logger->expects($this->never())
             ->method('log');
 
-        $out = $this->command->delete($site_name);
+        $out = $this->command->delete($this->site_name);
         $this->assertNull($out);
     }
 
@@ -72,10 +148,10 @@ class DeleteCommandTest extends CommandTestCase
      */
     public function testDeleteFailure()
     {
-        $site_name = 'my-site';
         $exception_message = 'Error message';
 
         $this->expectConfirmation();
+        $this->expectWorkflowProcessing();
         $this->site->expects($this->once())
             ->method('delete')
             ->with()
@@ -85,7 +161,7 @@ class DeleteCommandTest extends CommandTestCase
 
         $this->setExpectedException(\Exception::class, $exception_message);
 
-        $out = $this->command->delete($site_name);
+        $out = $this->command->delete($this->site_name);
         $this->assertNull($out);
     }
 }